refactor: 역할에 맞게 패키지 변경
