+++ tomcat/src/main/java/org/apache/catalina/controller/RequestMapping.java
@@ -1,10 +1,9 @@
-package org.apache.coyote.http11;
+package org.apache.catalina.controller;
 
 import com.techcourse.controller.LoginController;
 import com.techcourse.controller.RegisterController;
 import com.techcourse.controller.RootController;
 import com.techcourse.controller.StaticResourceController;
-import org.apache.catalina.controller.Controller;
 import org.apache.coyote.http11.request.HttpRequest;
 
 public class RequestMapping {
